Created calculateFare sub
Added overflow option for fare calculation. If the GPS can't find a road
path to the destination, the internal variable for
CALCULATE_TRAVEL_DISTANCE_BETWEEN_POINTS returns 100000m. If this is the
case, the fallback option becomes the straight-line distance between the
origin and destination. In most cases, this shouldn't reduce the fare by
too much.